ctfs/tsgctf-2024/CONPASS/requirements.txt
pycryptodome
fastapi
uvicorn
pydantic